requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py39h7f8727e_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py39h06a4308_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
bottleneck=1.3.4=py39hce1f21e_0
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.4.26=h06a4308_0
certifi=2021.10.8=py39h06a4308_2
cffi=1.15.0=py39hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.4=pyh9f0ad1d_0
cryptography=36.0.0=py39h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py39h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py39h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.4.4=h295c915_0
ffmpeg=4.3=hf484d3e_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
giflib=5.2.1=h7b6447c_0
glib=2.56.2=hd408876_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
icu=58.2=he6710b0_3
idna=3.3=pyhd3eb1b0_0
intel-openmp=2022.0.1=h06a4308_3633
ipykernel=6.9.1=py39h06a4308_0
ipython=8.2.0=py39h06a4308_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py39h06a4308_1
jinja2=3.0.3=pyhd3eb1b0_0
joblib=0.17.0=py_0
jpeg=9d=h7f8727e_0
jsonschema=4.4.0=py39h06a4308_0
jupyter=1.0.0=py39h06a4308_7
jupyter_client=7.1.2=pyhd3eb1b0_0
jupyter_console=6.4.3=pyhd3eb1b0_0
jupyter_core=4.9.2=py39h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.3.2=py39h295c915_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.3.0=hdf63c60_0
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
lz4-c=1.9.3=h295c915_1
markupsafe=2.0.1=py39h27cfd23_0
matplotlib=3.5.1=py39h06a4308_1
matplotlib-base=3.5.1=py39ha18d171_1
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py39h27cfd23_1000
mkl=2020.2=256
mkl-service=2.3.0=py39he8ac12f_0
mkl_fft=1.3.0=py39h54f3939_0
mkl_random=1.0.2=py39h63df603_0
munkres=1.1.4=py_0
nbclient=0.5.13=py39h06a4308_0
nbconvert=6.4.4=py39h06a4308_0
nbformat=5.3.0=py39h06a4308_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.5=py39h06a4308_0
nettle=3.7.3=hbbd107a_1
notebook=6.4.8=py39h06a4308_0
numexpr=2.7.3=py39hb2eb853_0
numpy=1.19.2=py39h89c1606_0
numpy-base=1.19.2=py39h2ae0177_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1h=h7b6447c_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.1=py39h295c915_1
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.0.1=py39h22f2fdc_0
pip=21.2.4=py39h06a4308_0
prometheus_client=0.13.1=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py39h2531618_6
pyrsistent=0.18.0=py39heee7806_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.0=hdb3f193_2
python-dateutil=2.8.1=py_0
python-fastjsonschema=2.15.1=pyhd3eb1b0_0
pytorch=1.11.0=py3.9_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2020.1=py_0
pyzmq=22.3.0=py39h295c915_2
qt=5.9.7=h5867ecd_1
qtconsole=5.3.0=pyhd3eb1b0_0
qtpy=2.0.1=pyhd3eb1b0_0
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd3eb1b0_0
scikit-learn=1.0.2=py39h51133e4_1
scipy=1.6.2=py39h91f5cce_0
seaborn=0.11.2=pyhd3eb1b0_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=61.2.0=py39h06a4308_0
sip=4.19.13=py39h295c915_0
six=1.15.0=py_0
soupsieve=2.3.1=pyhd3eb1b0_0
sqlite=3.38.2=hc218d9a_0
stack_data=0.2.0=pyhd3eb1b0_0
terminado=0.13.1=py39h06a4308_0
testpath=0.5.0=pyhd3eb1b0_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.11=h1ccaba5_0
torchaudio=0.11.0=py39_cu113
torchdata=0.3.0=pyhd8ed1ab_0
torchtext=0.12.0=py39
torchvision=0.12.0=py39_cu113
tornado=6.1=py39h27cfd23_0
tqdm=4.64.0=pyhd8ed1ab_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.8=pyhd3eb1b0_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py39h06a4308_1
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py39h06a4308_0
xz=5.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zlib=1.2.12=h7f8727e_1
zstd=1.4.9=haebb681_0
